MinMinBE/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blinker==1.9.0
boto3==1.39.8
botocore==1.39.8
CacheControl==0.14.2
cachetools==5.5.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==7.0.0
Django==5.1.3
django-admin-volt==1.0.14
django-allauth==65.2.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-leaflet==0.32.0
django-otp==1.5.4
django-seed==0.3.1
django-timezone-field==7.1
django-unfold==0.50.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework-gis==1.2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exponent_server_sdk==2.1.0
Faker==33.1.0
firebase-admin==6.6.0
Flask==3.1.1
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.4.2
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.13.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
toposort==1.10
twilio==9.3.7
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.8.2
yarl==1.18.0
zope.interface==7.2
